Remove callable trait
It's now only implemented once so having a trait is unnecessary.
