+++ src/compiler/typescript_to_rust/azle_generate/src/traits/mod.rs
@@ -1,4 +1,3 @@
-pub use callable::Callable;
 pub use get_name::GetName;
 pub use get_source_file_info::GetSourceFileInfo;
 pub use get_source_info::GetSourceInfo;
@@ -7,7 +6,6 @@ pub use get_span::GetSpan;
 pub use get_ts_type::GetTsType;
 pub use type_to_string::TypeToString;
 
-mod callable;
 mod get_name;
 mod get_source_file_info;
 mod get_source_info;
